pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["zhonghui <liaozhonghui@playdayy.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
httpx = "^0.28.0"
pydantic = "^2.10.2"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
sqlmodel = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.389"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
